Moved CookiePlaceholder story into the VideoEmbed Story to reduce story clutter.
