+++ base.scss
@@ -102,7 +102,6 @@
 @import "./src/stories/Library/article-header/article-header";
 @import "./src/stories/Library/content-list/content-list";
 @import "./src/stories/Library/content-list-item/content-list-item";
-@import "./src/stories/Library/cookie-placeholder/cookie-placeholder";
 @import "./src/stories/Library/image-credited/image-credited";
 @import "./src/stories/Library/event-description/event-description";
 @import "./src/stories/Library/link-with-icon/link-with-icon";
@@ -112,6 +111,7 @@
 @import "./src/stories/Library/card/card";
 @import "./src/stories/Library/paragraphs/paragraphs";
 @import "./src/stories/Library/video-embed/video-embed";
+@import "./src/stories/Library/video-embed/cookie-placeholder";
 @import "./src/stories/Library/card-grid/card-grid";
 @import "./src/stories/Library/promo-title/promo-title";
 @import "./src/stories/Library/Modals/group-modal-item-skeleton";
