+++ src/stories/Library/video-embed/cookie-placeholder.scss
@@ -28,6 +28,6 @@
   margin-bottom: $s-lg;
 }
 
-.cookie-placeholder__hide{
+.cookie-placeholder__hide {
   display: none;
 }
